+++ example/hello-rsp/webapp/main.dart
@@ -3,8 +3,12 @@ library hello_rsp;
 
 import "dart:async";
 import "package:stream/stream.dart";
+import "package:rikulo_commons/logging.dart";
+import "package:logging/logging.dart" show Logger, Level;
 
 part "helloView.rsp.dart"; //generated from helloView.rsp.html
+
+final logger = Logger('example');
 
 //URI mapping
 var _mapping = {
@@ -12,5 +16,8 @@ var _mapping = {
 };
 
 void main() {
+  Logger.root.level = Level.INFO;
+  logger.onRecord.listen(simpleLoggerHandler);
+
   StreamServer(uriMapping: _mapping).start();
 }