+++ lib/stream.dart
@@ -8,7 +8,7 @@ import "dart:io";
 import "dart:async";
 import "dart:collection" show HashMap;
 import "dart:convert" as cvt;
-import "package:logging/logging.dart" show Logger, Level;
+import "package:logging/logging.dart" show Logger;
 import "package:path/path.dart" as Path;
 import "package:charcode/ascii.dart";
 
@@ -16,7 +16,6 @@ import "package:rikulo_commons/util.dart";
 import "package:rikulo_commons/io.dart";
 import "package:rikulo_commons/convert.dart";
 import "package:rikulo_commons/browser.dart";
-import "package:rikulo_commons/logging.dart";
 
 import "plugin.dart";
 export "plugin.dart" show Router, DefaultRouter;
@@ -29,3 +28,5 @@ part "src/server.dart";
 part "src/connect_impl.dart";
 part "src/server_impl.dart";
 part "src/rsp_util.dart";
+
+final _logger = Logger('stream');