+++ test/issue29/webapp/main.dart
@@ -3,7 +3,14 @@
 // Author: tomyeh
 
 import "package:stream/stream.dart";
+import "package:rikulo_commons/logging.dart";
+import "package:logging/logging.dart" show Logger, Level;
+
+final logger = Logger('test');
 
 void main() {
+  Logger.root.level = Level.INFO;
+  logger.onRecord.listen(simpleLoggerHandler);
+
   StreamServer(homeDir: "static").start();
 }